requirements.txt
greenlet==1.1.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
